requirements.txt
phidata>=2.0.0
ollama>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.0